Refactor a little and split up the implementation of SMW to help with re-use Add new ObjectMapWidget and begin its implementation Add LevelEditor beginnings to SMW
